more tests, cleanup of structure
