+++ src/performance-log-middleware/IOptions.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.Extensions.Logging;
+
+namespace PerformanceLog
+{
+    public interface IOptions
+    {
+        IOptions WithLogLevel(LogLevel logLevel);
+        IOptions WithFormat(string format);
+    }
+}
