+++ src/performance-log-middleware/IPerformanceLogOptions.cs
@@ -1,0 +1,8 @@
+﻿namespace PerformanceLog
+{
+    public interface IPerformanceLogOptions
+    {
+        IOptions Configure();
+        void Default();
+    }
+}
