+++ src/performance-log-middleware/LogItem.cs
@@ -1,0 +1,15 @@
+﻿namespace PerformanceLog
+{
+    public class LogItem
+    {
+        public long Duration { get; set; }
+        public string Operation { get; set; }
+
+        public string CorrelationId { get; set; }
+
+        public override string ToString()
+        {
+            return $"Operation: {Operation}; Duration: {Duration}; CorrelationId: {CorrelationId}";
+        }
+    }
+}
