+++ src/performance-log-middleware/PerformanceLogMiddleware.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
-using Microsoft.AspNetCore.Builder;
 using System.Diagnostics;
 
 namespace PerformanceLog
@@ -63,95 +61,4 @@ namespace PerformanceLog
 
         }
     }
-
-    public static class PerformanceLogMiddlewareExtension
-    {
-        public static IApplicationBuilder UsePerformanceLog(this IApplicationBuilder app, PerformanceLogOptions options)
-        {
-            if (app == null)
-                throw new ArgumentNullException(nameof(app));
-            if (options == null)
-                throw new ArgumentNullException(nameof(options));
-
-            return app.UseMiddleware<PerformanceLogMiddleware>(options);
-        }
-
-        public static IApplicationBuilder UsePerformanceLog(this IApplicationBuilder app, Action<IPerformanceLogOptions> optionsAction)
-        {
-            if (app == null)
-                throw new ArgumentNullException(nameof(app));
-            if (optionsAction == null)
-                throw new ArgumentNullException(nameof(optionsAction));
-            var options = new PerformanceLogOptions();
-
-            optionsAction.Invoke(options);
-
-            return app.UseMiddleware<PerformanceLogMiddleware>(options);
-        }
-    }
-
-    public interface IPerformanceLogOptions
-    {
-        IOptions Configure();
-        void Default();
-    }
-
-    public interface IOptions
-    {
-        IOptions WithLogLevel(LogLevel logLevel);
-        IOptions WithFormat(string format);
-    }
-
-    public class PerformanceLogOptions : IPerformanceLogOptions, IOptions
-    {
-        public LogLevel LogLevel { get; set; }
-
-        public string Format { get; set; }
-
-        public PerformanceLogOptions()
-        {
-            Default();
-        }
-
-        public void Default()
-        {
-            LogLevel = LogLevel.Information;
-            Format = "request to {Operation} took {Duration}ms";
-        }
-
-        public IOptions WithFormat(string format)
-        {
-            this.Format = format;
-            return this;
-        }
-
-        public IOptions Configure()
-        {
-            return this;
-        }
-
-        public IOptions WithLogLevel(LogLevel logLevel)
-        {
-            this.LogLevel = logLevel;
-            return this;
-        }
-
-    }
-
-
-
-    public class LogItem
-    {
-        public long Duration { get; set; }
-        public string Operation { get; set; }
-
-        public string CorrelationId { get; set; }
-
-        public override string ToString()
-        {
-            return $"Operation: {Operation}; Duration: {Duration}; CorrelationId: {CorrelationId}";
-        }
-    }
-
-
 }
